+++ VirtualShop.vcxproj
@@ -141,6 +141,9 @@
   <ItemGroup>
     <ClCompile Include="Account.h" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="FileData.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
